+++ pymc3_models/__init__.py
@@ -1,3 +1,4 @@
 from pymc3_models.models.HierarchicalLogisticRegression import HierarchicalLogisticRegression
 from pymc3_models.models.LinearRegression import LinearRegression
 from pymc3_models.models.LogisticRegression import LogisticRegression
+from pymc3_models.models.NaiveBayes import GaussianNaiveBayes